refactor: 카드 별칭 업데이트 mockAPI를 통해서 동작하도록 개선
